ref(ts): Convert sessionHistory to ts
